Rename ov5_021DB888 to vs_seeker
