+++ platinum.us/main.lsf
@@ -479,7 +479,7 @@ Overlay overlay5
 	Object main.nef.p/src_overlay005_ov5_021D5BC0.c.o
 	Object main.nef.p/src_overlay005_ov5_021D5CB0.c.o
 	Object main.nef.p/src_overlay005_ov5_021D5EB8.c.o
-	Object main.nef.p/src_overlay005_ov5_021DB888.c.o
+	Object main.nef.p/src_overlay005_vs_seeker.c.o
 	Object main.nef.p/src_overlay005_ov5_021DC018.c.o
 	Object main.nef.p/src_overlay005_ov5_021DD42C.c.o
 	Object main.nef.p/src_overlay005_ov5_021DD6FC.c.o
