+++ src/overlay005/vs_seeker.c
@@ -24,7 +24,7 @@
 #include "unk_0206A8DC.h"
 #include "unk_0206AFE0.h"
 #include "unk_0207D3B8.h"
-#include "overlay005/ov5_021DB888.h"
+#include "overlay005/vs_seeker.h"
 #include "overlay005/ov5_021DFB54.h"
 
 typedef struct {
